cleanups and file renames
